Created a Projects Page and Cards for this.
Created a Projects Page and Cards for this.

File(s) created:
- /src/app/components/*Cards.jsx*
- /src/app/components/*Projects.jsx*
- /public/images/projects/*1.png*
- /public/images/projects/*2.png*
- /public//images/*about.jpg*

File(s) modified:
- /src/app/*page.js*
- /src/app/components/*Hero.jsx*
- /src/app/components/*About.jsx*
- /src/app/components/*Projects.jsx*
- /src/app/components/*Cards.jsx*

File(s) Removed:
- N/A

Detailed Overview of the changes made:

> [page.js]    - Imported and added Projects to the main landing page.

> [Hero.jsx]    - Placeholder image has been added (huh cat).

> [About.jsx]   - I have corrected the tab switching so the useState loads "Skills" rather than "Experience" on load.
                - Placeholder image has been added (another huh cat).

> [Cards.jsx]   - Started the development of Cards (please see note).

> [Projects.jsx]    - Created project_DATA with identifiers for my projects.
                    - Added the div for the information and Cards.jsx component (again, see notes).

Note that the site is currently not functional in this state with cards being incomplete and I have uploaded this as I am on the go and need the latest changed made from my Desktop at home. Next production deployment will be when this has been resolved.